feat: добавить работу с БД, возможность бронировать вещи и оставлять к ним отзывы

+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface UserService {
     User addNew(User user);
 
-    User update(User user, int userId);
+    User update(User user, long userId);
 
-    User getById(int userId);
+    User getById(long userId);
 
-    void delete(int userId);
+    void delete(long userId);
 
     List<User> getAll();
 }
